refactor(server): extract profile + PIN routes (6/8 server split)
Moves the 5 user-profile / PIN / resume-text endpoints to their own
Blueprint:
- GET  /api/profile      (open; pin_hash stripped server-side)
- POST /api/profile      (Bearer-gated; updates preferences + default resume)
- POST /api/resume       (resume-text upload → skill extraction)
- POST /api/verify-pin   (open)
- POST /api/set-pin      (Bearer-gated)

All write paths gate via check_api_secret(request) from core.config.

Live-lookup idiom for DB_PATH:
  from core import config as _config
  # ...later: _config.DB_PATH

Tests reach into the route layer via monkeypatch.setattr on
core.config.DB_PATH (and the existing patches on server.DB_PATH +
profile_manager.DB_PATH stay in place). Both client + secured_client
fixtures in test_profile_default_resume.py picked up an explicit
monkeypatch line so the blueprint sees the temp DB.

server.py: 675 → 577 LOC. -98 LOC. Two extractions to go (applications
+ resume versions); both are mechanical now that the patterns are in
place.

Tests: 219 passed (unchanged).
