+++ backend/server.py
@@ -86,6 +86,11 @@ app.register_blueprint(data_bp)
 # POST /api/scrape, GET /api/scrape/status.
 from routes.scrape_routes import scrape_bp
 app.register_blueprint(scrape_bp)
+# Profile + PIN + resume-text endpoints extracted to routes/profile_routes.py
+# (PR 6/8). Five routes: /api/profile (GET/POST), /api/resume, /api/verify-pin,
+# /api/set-pin. All write paths Bearer-gated when API_SECRET is set.
+from routes.profile_routes import profile_bp
+app.register_blueprint(profile_bp)
 
 
 # State + state singleton are imported above from core.state (PR 2/8).
@@ -113,93 +118,7 @@ def _check_api_secret() -> bool:
 
 # ─── Profile & Resume endpoints ────────────────────────────────────
 
-@app.route("/api/profile", methods=["GET"])
-def api_get_profile():
-    """Get user profile (preferences + default_resume_version).
-
-    Auth: deliberately left open. The payload is the user's own
-    preferences (custom skills, dream-company list, default resume key) —
-    not credentials. The pin_hash is stripped in ``get_profile`` so
-    readers can't even mount an offline brute-force. If a deployment
-    decides this should be gated, flip the same Bearer check used on POST.
-    """
-    try:
-        from storage.profile_manager import init_profile_tables, get_profile
-        init_profile_tables(DB_PATH)
-        return jsonify(get_profile(DB_PATH)), 200, {"Access-Control-Allow-Origin": "*"}
-    except Exception as e:
-        return jsonify({"error": str(e)}), 500
-
-
-@app.route("/api/profile", methods=["POST", "OPTIONS"])
-def api_update_profile():
-    """Update preferences (custom_skills, dream_*, default_resume_version).
-
-    Auth: Bearer-gated when API_SECRET is configured. Without this, an
-    anonymous attacker could flip ``default_resume_version`` and silently
-    change which resume the dashboard auto-scores every job against
-    (Issue #39 part A R2 #1). OPTIONS preflights pass through so browser
-    CORS preflight isn't blocked by the missing Authorization header.
-    """
-    if request.method == "OPTIONS":
-        return "", 204
-    if not _check_api_secret():
-        return jsonify({"error": "unauthorized"}), 401, {"Access-Control-Allow-Origin": "*"}
-    try:
-        from storage.profile_manager import (
-            init_profile_tables,
-            update_profile,
-            ProfileValidationError,
-        )
-        init_profile_tables(DB_PATH)
-        try:
-            update_profile(request.get_json() or {}, DB_PATH)
-        except ProfileValidationError as ve:
-            # Validation failures (e.g. unknown default_resume_version key)
-            # are client errors → 400 lets the dashboard show a useful
-            # message instead of an opaque 500.
-            return jsonify({"error": str(ve)}), 400, {"Access-Control-Allow-Origin": "*"}
-        return jsonify({"status": "updated"}), 200, {"Access-Control-Allow-Origin": "*"}
-    except Exception as e:
-        return jsonify({"error": str(e)}), 500
-
-
-@app.route("/api/resume", methods=["POST", "OPTIONS"])
-def api_upload_resume():
-    """Upload resume text → extract skills → store in DB."""
-    if request.method == "OPTIONS":
-        return "", 204
-    try:
-        from storage.profile_manager import init_profile_tables, upload_resume
-        init_profile_tables(DB_PATH)
-        data = request.get_json() or {}
-        resume_text = data.get("resume_text", "").strip()
-        if not resume_text:
-            return jsonify({"error": "resume_text field required"}), 400
-        skills = upload_resume(resume_text, DB_PATH)
-        return jsonify({
-            "status": "ok",
-            "skills_extracted": len(skills),
-            "skills": skills,
-        }), 200, {"Access-Control-Allow-Origin": "*"}
-    except Exception as e:
-        return jsonify({"error": str(e)}), 500
-
-
-@app.route("/api/verify-pin", methods=["POST", "OPTIONS"])
-def api_verify_pin():
-    if request.method == "OPTIONS":
-        return "", 204
-    try:
-        from storage.profile_manager import init_profile_tables, verify_pin
-        init_profile_tables(DB_PATH)
-        pin = (request.get_json() or {}).get("pin", "")
-        return jsonify({"verified": verify_pin(pin, DB_PATH)}), 200, {
-            "Access-Control-Allow-Origin": "*"
-        }
-    except Exception as e:
-        return jsonify({"error": str(e)}), 500
-
+# /api/profile, /api/resume, /api/verify-pin extracted to routes/profile_routes.py (PR 6/8).
 
 # ─── Application tracker endpoints ───────────────────────────────
 
@@ -609,24 +528,7 @@ def api_delete_resume_version(version_key):
         {"Access-Control-Allow-Origin": "*"}
 
 
-@app.route("/api/set-pin", methods=["POST", "OPTIONS"])
-def api_set_pin():
-    if request.method == "OPTIONS":
-        return "", 204
-    if not _check_api_secret():
-        return jsonify({"error": "unauthorized"}), 401
-    try:
-        from storage.profile_manager import init_profile_tables, set_pin
-        init_profile_tables(DB_PATH)
-        pin = (request.get_json() or {}).get("pin", "")
-        if len(pin) < 4:
-            return jsonify({"error": "PIN must be at least 4 characters"}), 400
-        set_pin(pin, DB_PATH)
-        return jsonify({"status": "pin_set"}), 200, {"Access-Control-Allow-Origin": "*"}
-    except Exception as e:
-        return jsonify({"error": str(e)}), 500
-
-
+# /api/set-pin extracted to routes/profile_routes.py (PR 6/8).
 # / (self-describing root) extracted to routes/data_routes.py (PR 4/8).
 
 # ─── CORS preflight ───────────────────────────────────────────────
